# backend/api/article/spec/openapi.yaml
openapi: 3.0.0
info:
  title: article api
  description: |
    tamaco-blog 記事管理システムのバックエンドAPIサービス。
    ブログ記事のCRUD操作、カテゴリ・タグの管理、
    記事関連の各種操作を提供するRESTful APIです。
  version: 1.0.0
servers:
  - url: http://localhost:{port}/article/
    description: dev
    variables:
      port:
        default: '8080'
  - url: https://api.<stg-domain-name>/article/
    description: stg
  - url: https://api.<prd-domain-name>/article/
    description: prd
security:
  - bearerAuth: []
paths:
...
  /v1/articles:
    get:
      tags:
        - articles
      summary: 記事一覧取得
      description: |
        記事の一覧を取得します。
        ページネーション、フィルタリング、ソートに対応しています。
      operationId: getArticles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/CategoryParam'
        - $ref: '#/components/parameters/TagParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: 記事一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
    post:
      tags:
        - articles
      summary: 記事作成
      description: 新しい記事を作成します。認証が必要です。
      operationId: createArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreate'
      responses:
        '201':
          description: 記事の作成に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '409':
          description: 同じIDの記事が既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyExists'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /v1/articles/{article_id}:
    get:
      tags:
        - articles
      summary: 記事詳細取得
      description: 記事IDを指定して記事の詳細を取得します。
      operationId: getArticleByID
      parameters:
        - name: article_id
          in: path
          required: true
          description: 記事ID
          schema:
            type: string
            example: how-to-use-golang
      responses:
        '200':
          description: 記事詳細の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '404':
          description: 記事が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
    put:
      tags:
        - articles
      summary: 記事更新
      description: 記事IDを指定して記事を更新します。認証が必要です。
      operationId: updateArticleByID
      parameters:
        - name: article_id
          in: path
          required: true
          description: 記事ID
          schema:
            type: string
            example: how-to-use-golang
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleUpdate'
      responses:
        '200':
          description: 記事の更新に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 記事が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '409':
          description: 同じIDの記事が既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyExists'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - articles
      summary: 記事削除
      description: 記事IDを指定して記事を削除します。認証が必要です。
      operationId: deleteArticleByID
      parameters:
        - name: article_id
          in: path
          required: true
          description: 記事ID
          schema:
            type: string
            example: how-to-use-golang
      responses:
        '204':
          description: 記事の削除に成功
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 記事が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /v1/articles/{article_id}/publish:
    patch:
      tags:
        - articles
      summary: 記事公開
      description: 記事IDを指定して記事を公開状態にします。認証が必要です。
      operationId: publishArticleByID
      parameters:
        - name: article_id
          in: path
          required: true
          description: 記事ID
          schema:
            type: string
            example: how-to-use-golang
      responses:
        '200':
          description: 記事の公開に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 記事が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /v1/articles/{article_id}/views:
    patch:
      tags:
        - articles
      summary: 閲覧数更新
      description: 記事IDを指定して記事の閲覧数を更新します。
      operationId: incrementArticleViewCount
      parameters:
        - name: article_id
          in: path
          required: true
          description: 記事ID
          schema:
            type: string
            example: how-to-use-golang
      responses:
        '200':
          description: 閲覧数の更新に成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  view_count:
                    type: integer
                    description: 更新後の閲覧数
                    example: 1235
        '404':
          description: 記事が見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
  /v1/articles/popular:
    get:
      tags:
        - articles
      summary: 人気記事一覧取得
      description: 閲覧数の多い記事一覧を取得します。
      operationId: getPopularArticles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 人気記事一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
  /v1/articles/recent:
    get:
      tags:
        - articles
      summary: 最新記事一覧取得
      description: 最新の記事一覧を取得します。
      operationId: getRecentArticles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 最新記事一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
  /v1/articles/archive/{year}/{month}:
    get:
      tags:
        - articles
      summary: 月別アーカイブ取得
      description: 指定された年月の記事一覧を取得します。
      operationId: getArticlesByMonth
      parameters:
        - $ref: '#/components/parameters/YearParam'
        - $ref: '#/components/parameters/MonthParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 月別記事一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
  /v1/categories:
    get:
      tags:
        - categories
      summary: カテゴリ一覧取得
      description: すべてのカテゴリの一覧を取得します。
      operationId: getCategories
      responses:
        '200':
          description: カテゴリ一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryList'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
    post:
      tags:
        - categories
      summary: カテゴリ作成
      description: 新しいカテゴリを作成します。認証が必要です。
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: カテゴリの作成に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '409':
          description: 同じIDのカテゴリが既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyExists'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /v1/categories/{category_id}:
    get:
      tags:
        - categories
      summary: カテゴリ詳細取得
      description: カテゴリIDを指定してカテゴリの詳細を取得します。
      operationId: getCategoryByID
      parameters:
        - name: category_id
          in: path
          required: true
          description: カテゴリID
          schema:
            type: string
            example: technology
      responses:
        '200':
          description: カテゴリ詳細の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: カテゴリが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
    put:
      tags:
        - categories
      summary: カテゴリ更新
      description: カテゴリIDを指定してカテゴリを更新します。認証が必要です。
      operationId: updateCategoryByID
      parameters:
        - name: category_id
          in: path
          required: true
          description: カテゴリID
          schema:
            type: string
            example: technology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: カテゴリの更新に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: カテゴリが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '409':
          description: 同じIDのカテゴリが既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyExists'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - categories
      summary: カテゴリ削除
      description: カテゴリIDを指定してカテゴリを削除します。認証が必要です。
      operationId: deleteCategoryByID
      parameters:
        - name: category_id
          in: path
          required: true
          description: カテゴリID
          schema:
            type: string
            example: technology
      responses:
        '204':
          description: カテゴリの削除に成功
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: カテゴリが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /v1/categories/{category_id}/articles:
    get:
      tags:
        - categories
      summary: カテゴリ別記事一覧取得
      description: 指定されたカテゴリの記事一覧を取得します。
      operationId: getArticlesByCategoryID
      parameters:
        - name: category_id
          in: path
          required: true
          description: カテゴリID
          schema:
            type: string
            example: technology
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: カテゴリ別記事一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '404':
          description: カテゴリが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
  /v1/tags:
    get:
      tags:
        - tags
      summary: タグ一覧取得
      description: すべてのタグの一覧を取得します。
      operationId: getTags
      responses:
        '200':
          description: タグ一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagList'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
    post:
      tags:
        - tags
      summary: タグ作成
      description: 新しいタグを作成します。認証が必要です。
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: タグの作成に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '409':
          description: 同じIDのタグが既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyExists'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /v1/tags/{tag_id}:
    get:
      tags:
        - tags
      summary: タグ詳細取得
      description: タグIDを指定してタグの詳細を取得します。
      operationId: getTagByID
      parameters:
        - name: tag_id
          in: path
          required: true
          description: タグID
          schema:
            type: string
            example: golang
      responses:
        '200':
          description: タグ詳細の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: タグが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
    put:
      tags:
        - tags
      summary: タグ更新
      description: タグIDを指定してタグを更新します。認証が必要です。
      operationId: updateTagByID
      parameters:
        - name: tag_id
          in: path
          required: true
          description: タグID
          schema:
            type: string
            example: golang
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdate'
      responses:
        '200':
          description: タグの更新に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: タグが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '409':
          description: 同じIDのタグが既に存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlreadyExists'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      tags:
        - tags
      summary: タグ削除
      description: タグIDを指定してタグを削除します。認証が必要です。
      operationId: deleteTagByID
      parameters:
        - name: tag_id
          in: path
          required: true
          description: タグID
          schema:
            type: string
            example: golang
      responses:
        '204':
          description: タグの削除に成功
        '401':
          description: 認証が必要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: タグが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /v1/tags/{tag_id}/articles:
    get:
      tags:
        - tags
      summary: タグ別記事一覧取得
      description: 指定されたタグの記事一覧を取得します。
      operationId: getArticlesByTagID
      parameters:
        - name: tag_id
          in: path
          required: true
          description: タグID
          schema:
            type: string
            example: golang
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
      responses:
        '200':
          description: タグ別記事一覧の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '404':
          description: タグが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
  /v1/search:
    get:
      tags:
        - search
      summary: 記事検索
      description: |
        キーワードを指定して記事を検索します。
        タイトル、内容、要約を対象に検索を行います。
      operationId: searchArticles
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/OrderParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/CategoryParam'
        - $ref: '#/components/parameters/TagParam'
      responses:
        '200':
          description: 検索結果の取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleList'
        '400':
          description: リクエストパラメータが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
      security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ArticleList:
      title: ArticleListResponse
      type: object
      description: 記事一覧レスポンス
      required:
        - articles
        - pagination
      properties:
        articles:
          type: array
          description: 記事のリスト
          items:
            $ref: '#/components/schemas/Article'
        pagination:
          type: object
          description: ページネーション情報
          required:
            - current_page
            - total_pages
            - total_count
            - page_size
          properties:
            current_page:
              type: integer
              description: 現在のページ番号
              minimum: 1
              example: 1
            total_pages:
              type: integer
              description: 総ページ数
              minimum: 0
              example: 5
            total_count:
              type: integer
              description: 総件数
              minimum: 0
              example: 100
            page_size:
              type: integer
              description: 1ページあたりの件数
              minimum: 1
              example: 20
            has_next:
              type: boolean
              description: 次のページがあるかどうか
              example: true
            has_prev:
              type: boolean
              description: 前のページがあるかどうか
              example: false
      example:
        articles:
          - id: 550e8400-e29b-41d4-a716-446655440000
            title: GoでRESTful APIを構築する方法
            slug: how-to-build-restful-api-with-go
            summary: Go言語を使用したRESTful APIの構築方法について解説します
            status: published
            view_count: 1234
            published_at: '2024-01-01T09:00:00Z'
            created_at: '2024-01-01T09:00:00Z'
            updated_at: '2024-01-02T10:30:00Z'
        pagination:
          current_page: 1
          total_pages: 5
          total_count: 100
          page_size: 20
          has_next: true
          has_prev: false
    BadRequest:
      title: BadRequest
      type: object
      description: リクエストが不正な場合のエラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: BAD_REQUEST
        message:
          type: string
          description: エラーメッセージ
          example: リクエストパラメータが不正です
        details:
          type: array
          description: エラーの詳細情報
          items:
            type: object
            properties:
              field:
                type: string
                description: エラーが発生したフィールド名
              reason:
                type: string
                description: エラーの理由
      example:
        code: BAD_REQUEST
        message: リクエストパラメータが不正です
        details:
          - field: title
            reason: タイトルは必須です
          - field: content
            reason: コンテンツは必須です
...
  parameters:
    PageParam:
      name: page
      in: query
      description: ページ番号
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    PageSizeParam:
      name: page_size
      in: query
      description: 1ページあたりの取得件数
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    SortParam:
      name: sort
      in: query
      description: ソート項目
      required: false
      schema:
        type: string
        enum:
          - created_at
          - updated_at
          - view_count
          - published_at
        default: created_at
      example: created_at
    OrderParam:
      name: order
      in: query
      description: ソート順
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
      example: desc
    StatusParam:
      name: status
      in: query
      description: 記事の公開状態でフィルタリング
      required: false
      schema:
        type: string
        enum:
          - draft
          - published
          - private
      example: published
    CategoryParam:
      name: category
      in: query
      description: カテゴリスラッグでフィルタリング
      required: false
      schema:
        type: string
      example: backend-development
    TagParam:
      name: tag
      in: query
      description: タグスラッグでフィルタリング
      required: false
      schema:
        type: string
      example: golang
    SearchParam:
      name: q
      in: query
      description: 検索キーワード
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 100
      example: RESTful API
    YearParam:
      name: year
      in: path
      description: 年（YYYY形式）
      required: true
      schema:
        type: integer
        minimum: 2020
        maximum: 2100
      example: 2024
...